Backend/engine/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
reportlab==4.2.2
matplotlib==3.9.0
numpy==2.0.0